coverage for archives.py
